IGAPP-655: Move search route
