+++ web-ts/src/components/__tests__/NothingFoundFeedbackBox.spec.tsx
@@ -2,9 +2,9 @@
 
 import React from 'react'
 import { NothingFoundFeedbackBox } from '../NothingFoundFeedbackBox'
-import { SEARCH_ROUTE } from '../../../../modules/app/route-configs/SearchRouteConfig'
+import { SEARCH_ROUTE } from '../../../../web/src/modules/app/route-configs/SearchRouteConfig'
 import { shallow } from 'enzyme'
-import createLocation from '../../../../createLocation'
+import createLocation from '../../../../web/src/createLocation'
 import TextButton from '../../../../modules/common/components/TextButton'
 
 jest.mock('api-client', () => ({
