+++ web-ts/src/routes/SearchPage.tsx
@@ -6,7 +6,7 @@ import SearchInput from '../../../modules/common/components/SearchInput'
 import { CategoriesMapModel, CategoryModel } from 'api-client'
 import CategoryList from '../../categories/components/CategoryList'
 import { withTranslation, type TFunction } from 'react-i18next'
-import type { StateType } from '../../../modules/app/StateType'
+import type { StateType } from '../../../web/src/modules/app/StateType'
 import SearchFeedback from '../components/SearchFeedback'
 import type { LocationState } from 'redux-first-router'
 import normalizeSearchString from '../../../modules/common/utils/normalizeSearchString'
