+++ web-ts/src/routes/__tests__/SearchPage.spec.tsx
@@ -6,11 +6,11 @@ import ConnectedSearchPage, { SearchPage } from '../SearchPage'
 import { CategoriesMapModel, CategoryModel } from 'api-client'
 import { mount, shallow } from 'enzyme'
 import configureMockStore from 'redux-mock-store'
-import { SEARCH_ROUTE } from '../../../../modules/app/route-configs/SearchRouteConfig'
+import { SEARCH_ROUTE } from '../../../../web/src/modules/app/route-configs/SearchRouteConfig'
 import moment from 'moment'
-import createLocation from '../../../../createLocation'
-import CityModelBuilder from 'api-client/src/testing/CityModelBuilder'
-import theme from '../../../../modules/theme/constants/theme'
+import createLocation from '../../../../web/src/createLocation'
+import CityModelBuilder from '../api-client/src/testing/CityModelBuilder'
+import theme from '../../../../web/src/modules/theme/constants/theme'
 import { ThemeProvider } from 'styled-components'
 
 jest.mock('react-i18next')
